Fix time unit embedding generation
