+++ community/fm-asr-streaming-rag/chain-server/chains.py
@@ -21,8 +21,8 @@ from langchain.docstore.document import Document
 
 from accumulator import TextAccumulator
 from retriever import NVRetriever
-from common import get_logger, LLMConfig, TimeResponse, UserIntent
-from utils import get_llm, classify
+from common import get_logger, LLMConfig
+from utils import get_llm, classify, TimeResponse, UserIntent
 from prompts import RAG_PROMPT, INTENT_PROMPT, RECENCY_PROMPT, SUMMARIZATION_PROMPT
 
 logger = get_logger(__name__)
